requirements.txt
fastapi
uvicorn
motor
pymongo
passlib[bcrypt]
python-jose
pydantic
